+++ RandomTastes/Framework/ModData.cs
@@ -1,0 +1,9 @@
+﻿namespace RandomTastes.Framework
+{
+    public class ModData
+    {
+        public ModDataSave[] saves = { };
+
+        public ModData() { }
+    }
+}
